chore: move sponsors script
